Added a superclass and connections for controllers
There's now a couple hashmaps in the main application class that get populated with the panes and controllers for each class. The main controller gets references to both, while the rest of the controllers only get a reference to the controllers. This makes it possible to call methods from one controller to another. This reduces duplicate code by allowing us to create a new game from the title screen, for example.

Minor stuff:
    Added some more event handlers as well as some of the actual code inside them.
    Added code to load the settings screen
    Tweaked the layout of the settings screen
    A couple comments to help explain how things are laid out. Comments are not complete yet, but are hopefully enough to get an idea of what's going on.