Added bindings to make drawType selector work
Minor changes were also added to how the game redraws the board. This allows the board to be reset between drawing new cards, which is important to not have cards duplicated or trying to draw over other cards.

Additionally, the drawCards has been changed to be a wrapper to two other methods, drawTop and drawStacks. This method first resets the board, then calls the other two methods. This new method does not take any other arguments, instead passing the 'newGame' field of the current GameController